refactor helpers and parser packages
