+++ tests/aa_pbs_exporter/resources/helpers_2.py
@@ -9,7 +9,7 @@ from pydantic.json import pydantic_encoder
 
 from aa_pbs_exporter.pbs_2022_01.models.parse_result import ParseResult
 from aa_pbs_exporter.pbs_2022_01.models.raw import IndexedString, ParsedIndexedString
-from aa_pbs_exporter.pbs_2022_01.parsers import IndexedStringParser
+from aa_pbs_exporter.pbs_2022_01.parser.parsers import IndexedStringParser
 from aa_pbs_exporter.snippets.file.validate_file_out import validate_file_out
 
 B = TypeVar("B", bound=ParsedIndexedString)
